+++ src/main/java/io/coodoo/workhorse/jobengine/control/BaseJobWorker.java
@@ -190,6 +190,18 @@ public abstract class BaseJobWorker {
     }
 
     /**
+     * Adds a timestamp followed by an info marker and the info message text in as a new line to the executions log<br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Info marker: Only if defined in {@link JobEngineConfig#LOG_INFO_MARKER}<br>
+     * Example: <code>[22:06:42.680] Step 3 complete</code>
+     * 
+     * @param message text to log
+     */
+    public void logInfo(String message) {
+        jobContext.logInfo(message);
+    }
+
+    /**
      * Adds a timestamp followed by an info marker and the info message text in as a new line to the executions log and also adds the message in severity INFO
      * to the server log<br>
      * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
@@ -204,15 +216,15 @@ public abstract class BaseJobWorker {
     }
 
     /**
-     * Adds a timestamp followed by an info marker and the info message text in as a new line to the executions log<br>
+     * Adds a timestamp followed by an warn marker and the warn message as a new line to the executions log<br>
      * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
-     * Info marker: Only if defined in {@link JobEngineConfig#LOG_INFO_MARKER}<br>
-     * Example: <code>[22:06:42.680] Step 3 complete</code>
+     * Error marker: <code>[WARN]</code> or as defined in {@link JobEngineConfig#LOG_WARN_MARKER}<br>
+     * Example: <code>[22:06:42.680] [WARN] Well thats suspicious...</code>
      * 
      * @param message text to log
      */
-    public void logInfo(String message) {
-        jobContext.logInfo(message);
+    public void logWarn(String message) {
+        jobContext.logWarn(message);
     }
 
     /**
@@ -230,15 +242,15 @@ public abstract class BaseJobWorker {
     }
 
     /**
-     * Adds a timestamp followed by an warn marker and the warn message as a new line to the executions log<br>
+     * Adds a timestamp followed by an error marker and the error message as a new line to the executions log<br>
      * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
-     * Error marker: <code>[WARN]</code> or as defined in {@link JobEngineConfig#LOG_WARN_MARKER}<br>
-     * Example: <code>[22:06:42.680] [WARN] Well thats suspicious...</code>
+     * Error marker: <code>[ERROR]</code> or as defined in {@link JobEngineConfig#LOG_ERROR_MARKER}<br>
+     * Example: <code>[22:06:42.680] [ERROR] Dafuq was that?!?!</code>
      * 
      * @param message text to log
      */
-    public void logWarn(String message) {
-        jobContext.logWarn(message);
+    public void logError(String message) {
+        jobContext.logError(message);
     }
 
     /**
@@ -256,15 +268,18 @@ public abstract class BaseJobWorker {
     }
 
     /**
-     * Adds a timestamp followed by an error marker and the error message as a new line to the executions log<br>
+     * Adds a timestamp followed by an error marker and the error message as a new line to the executions log. It also adds the message in severity ERROR and
+     * the throwable to the server log<br>
      * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
      * Error marker: <code>[ERROR]</code> or as defined in {@link JobEngineConfig#LOG_ERROR_MARKER}<br>
      * Example: <code>[22:06:42.680] [ERROR] Dafuq was that?!?!</code>
      * 
+     * @param logger server log logger
      * @param message text to log
+     * @param throwable cause of error
      */
-    public void logError(String message) {
-        jobContext.logError(message);
+    public void logError(Logger logger, String message, Throwable throwable) {
+        jobContext.logError(logger, message, throwable);
     }
 
     /**